+++ src/OrdersApi/Builder/APM/OxxoOrderBuilder.php
@@ -16,11 +16,6 @@ use Swag\PayPal\RestApi\V2\Api\Order\PaymentSource\Oxxo;
 
 class OxxoOrderBuilder extends AbstractAPMOrderBuilder
 {
-    public function isCompleteOnApproval(): bool
-    {
-        return true;
-    }
-
     protected function buildPaymentSource(
         AsyncPaymentTransactionStruct $paymentTransaction,
         SalesChannelContext $salesChannelContext,